requirements.txt
termcolor
dnspython
requests